requirements.txt
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==2.0.1
Flask-SQLAlchemy==2.4.3
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.5
grpcio-status==1.65.5
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
proto-plus==1.24.0
protobuf==5.27.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rsa==4.9
six==1.16.0
SQLAlchemy==1.3.8
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==2.0.1
